refactor: MyStudyData -> response로 변경
