+++ backend/src/main/java/com/woowacourse/moamoa/study/service/MyStudyService.java
@@ -2,10 +2,9 @@ package com.woowacourse.moamoa.study.service;
 
 import com.woowacourse.moamoa.common.exception.UnauthorizedException;
 import com.woowacourse.moamoa.member.domain.repository.MemberRepository;
-import com.woowacourse.moamoa.member.query.MemberDao;
 import com.woowacourse.moamoa.member.query.data.MemberData;
 import com.woowacourse.moamoa.study.query.MyStudyDao;
-import com.woowacourse.moamoa.study.query.data.MyStudyData;
+import com.woowacourse.moamoa.study.service.response.MyStudyResponse;
 import com.woowacourse.moamoa.study.query.data.MyStudySummaryData;
 import com.woowacourse.moamoa.study.service.response.MyStudiesResponse;
 import com.woowacourse.moamoa.tag.query.response.TagSummaryData;
@@ -45,15 +44,15 @@ public class MyStudyService {
         return new MyStudiesResponse(mapToResponse(myStudySummaryData, ownerWithStudyTags));
     }
 
-    private List<MyStudyData> mapToResponse(final List<MyStudySummaryData> myStudySummaryData,
-                                             final Map<Long, Map<MemberData, List<TagSummaryData>>> ownerWithStudyTags) {
+    private List<MyStudyResponse> mapToResponse(final List<MyStudySummaryData> myStudySummaryData,
+                                                final Map<Long, Map<MemberData, List<TagSummaryData>>> ownerWithStudyTags) {
 
-        List<MyStudyData> myStudyData = new ArrayList<>();
+        List<MyStudyResponse> myStudyData = new ArrayList<>();
         for (MyStudySummaryData studyData : myStudySummaryData) {
             final Map<MemberData, List<TagSummaryData>> ownerWithTag = ownerWithStudyTags.get(studyData.getId());
 
             for (MemberData owner : ownerWithTag.keySet()) {
-                myStudyData.add(new MyStudyData(studyData, owner, ownerWithTag.get(owner)));
+                myStudyData.add(new MyStudyResponse(studyData, owner, ownerWithTag.get(owner)));
             }
         }
         return myStudyData;
